# docs/openapi.yml
openapi: 3.0.3
info:
  title: CLAIM-TRACE API
  version: "1.0.0"
  description: |
    보험금 지급심사 근거 추적 및 인적 개입 기록 시스템의 REST API.

    ## 설계 목표
    - **G1** AI 권고와 사람 판정을 물리적으로 분리해 기록한다
    - **G2** 판단 근거를 항목 단위로, 채택된 것과 기각된 것 모두 남긴다
    - **G3** 근거를 내부용과 고객용 두 수준으로 분리해 노출한다
    - **G4** 인적 개입 조건을 정책으로 정의하고 그 이행을 강제한다

    ## 규제 대응 단일 진입점
    아래 세 엔드포인트에 불변조건 검증이 모인다. 우회 경로를 두지 않는다.

    | 엔드포인트 | 검증 |
    |---|---|
    | `POST /claim-items/{itemId}/reviews` | INV-1, INV-2, INV-3, INV-12 |
    | `POST /claims/{claimId}/decision` | INV-4, INV-10 |
    | `POST /claims/{claimId}/explanations/draft` | INV-5 |

    ## 공통 규칙
    - 날짜/시간 형식: ISO 8601
    - 목록 API는 `page`, `size` 쿼리 파라미터를 받고 `PagedResponse` 구조로 반환한다
    - 모든 오류 응답은 공통 `Error` 스키마를 사용한다
    - Security Schemes는 과제 평가 대상에서 제외되어 정의하지 않는다
servers:
  - url: http://localhost:8080/api
    description: 로컬 개발 서버
...
paths:
...
  /external/claims/{claimId}/recommendations:
    post:
      tags: [External]
      summary: AI 모델 권고 및 기여도 전달
      description: |
        AI 심사 모델이 항목별 보상제외 확률과 기여도를 전달한다.
        `ai_recommendations`와 `evidences`(source = AI)를 생성한다.
        권고 값(`recommendation`)은 확률과 임계값으로 계산되지 않으므로 모델이 직접 보낸다.

        **이 엔드포인트는 판정을 확정하지 않는다 (INV-13).**
        `reviews`, `interventions`, `claims.status`에 접근할 수 없다.
      operationId: submitRecommendations
      parameters:
        - $ref: '#/components/parameters/ClaimId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelName, modelVersion, threshold, items]
              properties:
                modelName: { type: string, example: claim-risk }
                modelVersion: { type: string, example: v2.3 }
                threshold: { type: number, format: double, example: 0.3 }
                items:
                  type: array
                  items:
                    type: object
                    required: [claimItemId, exclusionProbability, recommendation, contributions]
                    properties:
                      claimItemId: { type: integer, format: int64 }
                      exclusionProbability: { type: number, format: double }
                      recommendation: { $ref: '#/components/schemas/ItemDecision' }
                      contributions:
                        type: array
                        description: 항목별 새플리 기여도
                        items:
                          type: object
                          properties:
                            feature: { type: string, example: 동일 상병 30일 내 재청구 }
                            value: { type: number, format: double, example: 0.31 }
                            polarity: { $ref: '#/components/schemas/Polarity' }
      responses:
        '201':
          description: 권고 등록 완료
        '400': { $ref: '#/components/responses/BadRequest' }
        '403':
          description: 판정 확정 시도 등 허용되지 않은 쓰기 (E-11, INV-13)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
...
  parameters:
...
    ClaimId:
      name: claimId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: 요청 값 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: 대상을 찾을 수 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: OVERRIDE_REASON_REQUIRED }
        message: { type: string, example: AI 권고와 다른 판정에는 오버라이드 사유가 필요합니다 }
        invariant:
          type: string
          description: 위반된 불변조건 번호
          example: INV-3
        details:
          type: object
          description: 위반 상세. 오류 유형에 따라 구조가 다르다
        timestamp: { type: string, format: date-time }
...
    ItemDecision:
      type: string
      enum: [PAY, PARTIAL, DENY]
...
    Polarity:
      type: string
      enum: [POSITIVE, NEGATIVE]
...